Fixed the service worker script
An error was occurring when the service worker would interact
with the map API, this would cause the service worker to always look
for map files in the cache, causing errors when it couldn't find them.
Even when online. This has been fixed by checking the origin of the
request.